feat(traceql): MetricsPipeline lowering — rate/*_over_time → Aggregate(Scan(traces)) (RC2) (#153)
* feat(traceql): MetricsPipeline lowering — rate/*_over_time → Aggregate(Scan(traces)) (RC2)

Lower `RootExpr.MetricsPipeline` into chplan `Aggregate(<spanset-tree>)`:

- `rate()` / `count_over_time()` → `count(1)` (the wrapping
  RangeWindow.Func discriminates rate vs count_over_time at the
  handler layer).
- `sum_over_time(attr)` / `min_over_time` / `max_over_time` → CH
  `sum`/`min`/`max` of the attribute.
- `quantile_over_time(attr, q)` → `quantile(q)(attr)` (single-quantile
  only; multi-quantile surfaces a clean error).
- `by (label, ...)` → chplan Aggregate.GroupBy with parallel aliases.

Range/step intentionally NOT in the lowered tree — TraceQL doesn't
carry the range in the AST. The follow-up `/api/metrics/query_range`
handler wraps with `chplan.RangeWindow` at request time (see
docs/fork-tempo-plan.md § 2c).

Uses the tsouza/tempo:cerberus-accessors fork's `MetricsAggregate.{Op,
Attribute,GroupBy,Quantiles}` + `MetricsAggregate*` constants — zero
unsafe/reflect (post-#148 rule).

`avg_over_time(attr)` and `histogram_over_time(attr)` are deferred:
Tempo parses avg_over_time into an unexported `*averageOverTimeAggregator`
which the fork hasn't exposed accessors for yet, and histogram_over_time
needs a different chplan shape than scalar Aggregate.


* fix(api/loki): cast lineLimit to int64 for SelectBuilder.Limit

Origin/main currently fails `go build` because detected_fields.go
passes `lineLimit int` directly to `SelectBuilder.Limit(int64)`. The
build / lint required-check jobs are red on `main` (run 25790341717).
Cast at the call site so this PR's CI can go green. The fix is a
one-line type cast unrelated to the MetricsPipeline lowering work
this PR otherwise carries.


---------

+++ internal/api/loki/detected_fields.go
@@ -141,7 +141,7 @@ func buildDetectedFieldsSQL(s schema.Logs, matchers []*labels.Matcher, start, en
 		sb.Where(timeBoundFrag(s.TimestampColumn, "<=", end))
 	}
 	sb.OrderBy(chsql.Col(s.TimestampColumn), true).
-		Limit(lineLimit)
+		Limit(int64(lineLimit))
 
 	sqlStr, args := sb.Build()
 	return sqlStr, args, nil
